차지시간 변수 float로 바꿈

+++ JUMPMAN/app/src/main/java/kr/ac/kpu/game/s2017182016/jumpman/game/Player.java
@@ -40,8 +40,8 @@ public class Player implements GameObject, BoxCollidable {
     private double velocityX;
     private double jumpX;
     private double velocityY;
-    private int chargetime = 0;
-    private int prevchargetime = 0;
+    private float chargetime = 0;
+    private float prevchargetime = 0;
     private int isInverse = 1;
     private float ground_y;
     private float ground_y2;
